requirements.txt
# ============================================================================
# DEPENDENCIAS DEL PROYECTO - CLÍNICA DENTAL BACKEND
# ============================================================================
# Framework principal
Django==5.2.6
djangorestframework==3.16.1

# Base de datos PostgreSQL
psycopg2-binary==2.9.11
dj-database-url==2.1.0

# Multi-tenancy
django-tenants==3.9.0

# Autenticación JWT
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# CORS para frontend
django-cors-headers==4.7.0

# Filtros avanzados
django-filter==24.3

# Manejo de imágenes y archivos
Pillow==11.1.0

# Almacenamiento de archivos (AWS S3, etc.)
django-storages==1.14.6

# Variables de entorno
python-decouple==3.8

# Cache con Redis
django-redis==6.0.0
redis==5.0.1

# Notificaciones push (Firebase)
fcm-django==2.3.1

# Testing
pytest-django==4.11.1

# Servidor de producción
gunicorn==21.2.0
whitenoise==6.6.0

# Utilidades
python-dateutil==2.8.2
